Kafka data store: onSchemaDeleted: truncate topic records (do not delete topic)

+++ geomesa-kafka/geomesa-kafka-datastore/src/main/scala/org/locationtech/geomesa/kafka/data/KafkaDataStore.scala
@@ -278,7 +278,13 @@ class KafkaDataStore(
 
     WithClose(AdminClient.create(props)) { admin =>
       if (admin.listTopics().names().get.contains(topic)) {
-        admin.deleteTopics(Collections.singletonList(topic)).all().get
+        if (config.onSchemaDeleteTruncate) {
+          logger.info(s"Truncating topic $topic")
+          KafkaTruncateTopic(admin).truncate(topic)
+        } else {
+          logger.info(s"Deleting topic $topic")
+          admin.deleteTopics(Collections.singletonList(topic)).all().get
+        }
       } else {
         logger.warn(s"Topic [$topic] does not exist, can't delete it")
       }
@@ -513,6 +519,7 @@ object KafkaDataStore extends LazyLogging {
       consumers: ConsumerConfig,
       producers: ProducerConfig,
       clearOnStart: Boolean,
+      onSchemaDeleteTruncate: Boolean,
       topics: TopicConfig,
       @deprecated("unused")
       serialization: SerializationType,
